Move available upgrades into upgrades model
