--- js-src/modules/models/shipBuild/upgrades.js
@@ -1,60 +1,0 @@
-'use strict';
-
-var _ = require('lodash');
-var upgrades = require('../upgrades').keyed;
-
-module.exports = {
-    // Return array of upgrades of specific type which are legal to purchased for currernt build
-    //  (e.g. restricted by chassis, size, already a starting upgrade, already purchased etc.)
-    getFilteredUpgrades: function (upgradeType, build) {
-        var upgradesOfType = upgrades[upgradeType];
-        var existingUpgradesOfType = build.upgrades.allForType(upgradeType);
-
-        var filteredUpgrades = _.filter(upgradesOfType, function (upgrade) {
-            // Remove any upgrades for different ships
-            if (upgrade.ship && upgrade.ship.indexOf(build.currentShip.shipData.name) < 0) {
-                return false;
-            }
-
-            // Remove any upgrades for different ship sizes
-            if (upgrade.size && upgrade.size.indexOf(build.currentShip.shipData.size) < 0) {
-                return false;
-            }
-
-            // Don't show anything which is a starting upgrade for the ship
-            if (build.currentShip.startingUpgrades) {
-                var found = _.find(build.currentShip.startingUpgrades, function (startingUpgrade) {
-                    return startingUpgrade.xws === upgrade.xws;
-                });
-                if (found) {
-                    return false;
-                }
-            }
-
-            // Remove any upgrades the build already has
-            var upgradeExists = _.find(existingUpgradesOfType, function (existingUpgrade) {
-                return existingUpgrade.id === upgrade.id;
-            });
-
-            if (upgradeExists) {
-                var upgradeIsAllowed = false;
-                // filter out any upgrades the player already has
-                // except
-                // * secondary weapons & bombs
-                if (upgrade.slot === 'Bomb' || upgrade.slot === 'Torpedo' || upgrade.slot === 'Cannon' || upgrade.slot === 'Turret' || upgrade.slot === 'Missile') {
-                    upgradeIsAllowed = true;
-                // * hull upgrade and shield upgrade
-                } else if (upgrade.xws === 'hullupgrade' || upgrade.xws === 'shieldupgrade') {
-                    upgradeIsAllowed = true;
-                }
-                if (!upgradeIsAllowed) {
-                    return false;
-                }
-            }
-
-            return true;
-        });
-
-        return filteredUpgrades;
-    }
-};
